Stub out NativeFontMetrics for non-iOS platforms
