+++ packages/experimental/NativeFontMetrics/src/NativeFontMetrics.ios.ts
@@ -1,21 +1,7 @@
 import { NativeModules } from 'react-native';
+import { ScaleFactors, TextStyle } from './NativeFontMetrics.types';
 
 export const NativeFontMetrics = NativeModules.FRNFontMetrics;
-
-export type TextStyle =
-  | 'caption2'
-  | 'caption1'
-  | 'footnote'
-  | 'subheadline'
-  | 'callout'
-  | 'body'
-  | 'headline'
-  | 'title3'
-  | 'title2'
-  | 'title1'
-  | 'largeTitle';
-
-export type ScaleFactors = { [K in TextStyle]: number | undefined };
 
 interface NativeFontMetricsInterface {
   allScaleFactors(): ScaleFactors;
