+++ packages/experimental/NativeFontMetrics/src/NativeFontMetrics.types.ts
@@ -1,0 +1,14 @@
+export type TextStyle =
+  | 'caption2'
+  | 'caption1'
+  | 'footnote'
+  | 'subheadline'
+  | 'callout'
+  | 'body'
+  | 'headline'
+  | 'title3'
+  | 'title2'
+  | 'title1'
+  | 'largeTitle';
+
+export type ScaleFactors = { [K in TextStyle]?: number };
